+++ game/battle/processed_battle.lua
@@ -2118,6 +2118,7 @@ end
 
 --- Standby phase - player action phase
 -- Players can set animals, remove animals, move animals, or move biomes
+-- NOTE: This file is concatenated in build - do NOT return at file end
 local standby = function()
     local options = {'Set Animal', 'Move Animal', 'Remove Animal', 'Move Biome'}
     local actions = { _set_animal, _move_animal, _remove_animal, _move_biome }
@@ -2144,8 +2145,7 @@ local standby = function()
     _update_ui()
     return success
 end
-
-return standby
+-- DO NOT add 'return standby' - this file is concatenated in build process
 
 -- /home/s1eep1ess/workspace/lua/echeckers/game/battle/phases/3_battle_phase.lua
 -- Must be automatic,